+++ src/warehouse_ddd_petproject/domain/exceptions.py
