+++ src/warehouse_ddd_petproject/domain/model.py
